fix episode results by adding ID to encounters
