+++ src/Services/Qdm/Services/AbstractQdmService.php
@@ -41,6 +41,22 @@ abstract class AbstractQdmService
         $this->codeTypesService = $codeTypesService;
     }
 
+    public static function convertToObjectIdBSONFormat($id)
+    {
+        // max bigint size will fit in 16 characters so we will always have enough space for this.
+        $padded_hex = sprintf("%024X", $id);
+        return $padded_hex;
+    }
+
+    public static function convertIdFromBSONObjectIdFormat($id)
+    {
+        // max bigint size is 8 bytes which will fit fine
+        // string ID should be prefixed with 0s so the converted data type should be far smaller
+        $trimmedId = ltrim($id, '\x0');
+        $decimal = hexdec($trimmedId);
+        return $decimal;
+    }
+
     public function validDateOrNull($date)
     {
         if ($date == '0000-00-00') {
